OXY Plugin
CopperOxidation is a Minecraft plugin that adds functionality to copper blocks. When a player interacts with a copper block using a stick, the block's oxidation level will increase by one. This plugin includes a custom stick that can be obtained using the `/getoxidationstick` command.

This plugin is built using the Spigot API and is designed to be used on Spigot servers. It includes a command executor and event listener, and is easily customizable for server owners who want to modify the functionality of the plugin.